drivers: fih: Create procfs dir before creating cpu node
* This causes an annoying warning on almost every boot. On stock ROM AllHWList seems to already exist, before
fih_cpu_init is run, so proc_create has no problem creating the cpu node. However in custom ROMs this path is not
created, so we have to suffer through a crash before it is created properly. Address this by creating
the path before attempting to create the cpu node. Initial fix by @radu-v.

Change-Id: Ib031b5cd1a8443d6999e1940dd3fcbea52aa2366